worker/requirements.txt
# 단가표 통합 관리 시스템 - Worker 의존성

# Redis Queue
rq==1.16.0
redis==5.0.1

# 데이터베이스
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# 엑셀 파싱
openpyxl==3.1.2
pandas==2.2.0

# 이미지 처리 및 OCR
pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80

# 테이블 추출
img2table==1.2.11

# 텍스트 매칭
rapidfuzz==3.6.1

# 유틸리티
python-dotenv==1.0.1
